Improve register allocation.
The callee-saved registers are made available for allocation and are considered
after the caller-saved registers. Variables are still spilt to the stack when
all registers are exhausted.

48 bytes (40 bytes at 16-byte alignment) of each stack-frame are reserved to save
any of the callee-saved registers (rbx r12 r13 r14 r15), although only those
required are actually saved/restored. The prologue and epilogue functions have
been updated appropriately. Spilled variables are below this space.